fix vertex use async func to set auth creds
