+++ litellm/llms/fine_tuning_apis/vertex_ai.py
@@ -185,8 +185,14 @@ class VertexFineTuningAPI(VertexLLM):
             "creating fine tuning job, args= %s", create_fine_tuning_job_data
         )
 
+        _auth_header, vertex_project = self._ensure_access_token(
+            credentials=vertex_credentials,
+            project_id=vertex_project,
+        )
+
         auth_header, _ = self._get_token_and_url(
             model="",
+            auth_header=_auth_header,
             gemini_api_key=None,
             vertex_credentials=vertex_credentials,
             vertex_project=vertex_project,
@@ -251,8 +257,14 @@ class VertexFineTuningAPI(VertexLLM):
         vertex_credentials: str,
         request_route: str,
     ):
+
+        _auth_header, vertex_project = await self._ensure_access_token_async(
+            credentials=vertex_credentials,
+            project_id=vertex_project,
+        )
         auth_header, _ = self._get_token_and_url(
             model="",
+            auth_header=_auth_header,
             gemini_api_key=None,
             vertex_credentials=vertex_credentials,
             vertex_project=vertex_project,
