+++ litellm/llms/vertex_ai_and_google_ai_studio/gemini_embeddings/batch_embed_content_handler.py
@@ -43,8 +43,14 @@ class GoogleBatchEmbeddings(VertexLLM):
         client=None,
     ) -> EmbeddingResponse:
 
+        _auth_header, vertex_project = self._ensure_access_token(
+            credentials=vertex_credentials,
+            project_id=vertex_project,
+        )
+
         auth_header, url = self._get_token_and_url(
             model=model,
+            auth_header=_auth_header,
             gemini_api_key=api_key,
             vertex_project=vertex_project,
             vertex_location=vertex_location,
